# openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Sanctions Screening API",
    "description": "Screen entities against global sanctions lists (OFAC SDN, EU Consolidated List, UN Security Council) with explainable, auditable results.\n\nKey features:\n- 5-tier matching: exact, alias, fuzzy, phonetic, and LLM-assisted verification\n- LLM cascade reduces false positives on borderline matches\n- Complete audit trail for regulatory compliance\n- Human review queue for high-confidence hits\n- Per-tenant custom watchlists\n\nBuilt for fintech, VASP, and crypto exchange compliance teams.",
    "version": "0.10.0"
  },
  "servers": [
    {
      "url": "https://{rapidapi-host}",
      "description": "RapidAPI Hub (host provided per-request)"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for authentication. Every request must include a valid\n`X-API-Key` header. Keys are issued per tenant when the service runs in\ntenant mode (TENANT_MODE=1). In legacy deployments the key comes from the\n`API_KEY` environment variable."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "Machine-readable error code",
                "example": "MISSING_FIELD"
              },
              "message": {
                "type": "string",
                "description": "Human-readable error description",
                "example": "query_name is required"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "Case": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "tenant_id": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          },
          "query_name": {
            "type": "string"
          },
          "candidate": {
            "type": "string"
          },
          "list_name": {
            "type": "string"
          },
          "confidence_score": {
            "type": "number"
          },
          "llm_reasoning": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected",
              "false_positive"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "decided_at": {
            "type": "string",
            "format": "date-time"
          },
          "decided_by": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "query_name",
          "candidate",
          "status"
        ]
      },
      "CaseDecisionRequest": {
        "type": "object",
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "approved",
              "rejected",
              "false_positive"
            ]
          }
        },
        "required": [
          "decision"
        ]
      },
      "WatchlistVersionInfo": {
        "type": "object",
        "additionalProperties": {
          "type": "integer"
        },
        "description": "Map of list name -> current version."
      },
      "AuditEvent": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "tenant_id": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          },
          "event_type": {
            "type": "string",
            "enum": [
              "screening_requested",
              "screening_completed",
              "provider_error",
              "case_decision"
            ]
          },
          "payload_json": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "event_type",
          "payload_json"
        ]
      }
    }
  },
  "paths": {
...
    "/watchlists": {
      "get": {
        "summary": "List tenant watchlists and versions",
        "operationId": "listWatchlists",
        "tags": [
          "watchlists"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Watchlist names and current versions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "watchlists": {
                      "$ref": "#/components/schemas/WatchlistVersionInfo"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "API key required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/watchlists/{list_name}": {
      "post": {
        "summary": "Upload a tenant watchlist (CSV, new version)",
        "operationId": "uploadWatchlist",
        "tags": [
          "watchlists"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "list_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "text/csv": {
              "schema": {
                "type": "string"
              },
              "example": "entity_id,name,type,country,dob,identifiers,tags\nwl1,John Smith,person,US,1980-01-01,passport:123,pep\n"
            }
          }
        },
        "responses": {
          "200": {
            "description": "Uploaded (new version)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "list_name": {
                      "type": "string"
                    },
                    "version": {
                      "type": "integer"
                    },
                    "entries": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid CSV",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/cases": {
      "get": {
        "summary": "List review queue cases",
        "operationId": "listCases",
        "tags": [
          "review"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "approved",
                "rejected",
                "false_positive"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Case list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "cases": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Case"
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Insufficient tier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/cases/{id}": {
      "get": {
        "summary": "Get one review case",
        "operationId": "getCase",
        "tags": [
          "review"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Case detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Case"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Decide a pending review case",
        "operationId": "decideCase",
        "tags": [
          "review"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Decision recorded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "case_id": {
                      "type": "integer"
                    },
                    "decision": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Case not pending or not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/audit": {
      "get": {
        "summary": "Query the audit trail",
        "operationId": "listAudit",
        "tags": [
          "audit"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "event_type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "screening_requested",
                "screening_completed",
                "provider_error",
                "case_decision"
              ]
            }
          },
          {
            "name": "request_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Audit events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "events": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AuditEvent"
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Insufficient tier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  }
}